Commit Day 8
I had already done this but I didn't upload my code into github for some reason so i had to recreate the classes I had created for the cactuses. Once i have done that I got them to appear on the screen sequentially and move. Now i will add some more classes of like a rock or something and get them to repeat in that order and the guy will have to avoid them.