Commit Day 11
So the list idea seemed to not work as well as i had hoped for because I haven't been able to call out a function within a class using just the i from the iteration. Okay! So with your help i have been able to create a class to jam all of the obstacles into one class and then try to make one appear on the screen and reappear.  Okay I was able to do the first part but now the same issue with calling out a certain function is not able to happen.